RSE: SFCP: Allow receive message to be called multiple times
Previously, we could only call the sfcp_hal_receive_message function
once and in the case that it is called with a size smaller than the
total message size, the sender would hang. Allow the caller of
receive_message to instead call the function in a loop.

This allows for receiving the message into separate buffers and also
allows the receiver to drop the message safely while unblocking the
sender.

Change-Id: I3d9b902afdfc4d0863e87e54d6bcf771e479a4f1
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>