RSE: SFCP: Drop message on buffer allocation failure
Previously, in the interrupt handler, if the buffer fails to allocate
then we read out the message header only and use that to determine where
to send the error. Because we did not read out the whole message, the
sender would remain blocked waiting for us to ACK the message.

Add a new helper function which safely drops the message by reading it
out into a buffer provided by the caller (this can be a stack buffer).
Make use of this function in the interrupt handler to avoid the issue
described above.

Change-Id: Ifa9671f5afb110c500f0d48d1789ee7e8467600f
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>