RSE: Add support for CM/DM ROTPK revocation
Update the existing ROTPK revocation code to support specifying a
specific index to move the ROTPK revocation counter to.

We then update the existing non_endorsed_dm_provisioning_values
structure to follow the newly specified format. This allows for
specifying a specific index which can then be passed to the revocation
and update functions. Note this is not supported for the existing non
endorsed DM provisioning as they must always write new ROTPKs and cannot
only revoke them.

New functions are added to handle plain data messages for CM/DM
revocation; these will be used to handle the new authenticated plain
data format, with additional patches required to parse this format and
forward the message onto this handler.

Change-Id: I7778d3810c9a67b388578e9a787feeabbf6b0235
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>