RSE: Add authenticated plain data handler
Rework the existing blob handling code to also support a handler for the
authenticated plain data format.

This change refactors the blob parsing and validation code to take
in a pointer to the authentication header instead of the blob
itself. This allows the same validation code to
parse and validate an authenticated plain data message. The handler
source file name and header are renamed to reflect this change.

Note that the authenticated plain data handler does not contain all of
the same parsing as the blob, because many parts of the header are not
relevant for the authenticated plain data.

Change-Id: I8966941afe3a1baee1af7e5dfcee2fc4f91d3852
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>