RSE: SFCP: Add bus arbitration random delay
Add a new source file while contains the implementation for a basic LFSR
(taken from cc3xx_rng.c).

This is used within sfcp.c in the case that the packet fails to send due
to bus contention. In this case, we wait a random amount of time (up to
10000 cycles) before returning from the function. The intention is that
this will make it much more likely that the callers then try and send
messages at different time.

Note that there still must be logic in the caller to receive the message
if necessary instead of continue to send. Should the callers continue to
try and send their message, then bus contention will continue to occur.

Change-Id: I899a0188024ae6cf9551eab7ebb1f640ba47be98
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>