platform: include .TFM_BSS in zero table of GCC common linker script
While the copy table included the default data segment for these
templates, the zero table didn't include the default bss segment.

Make sure that area gets initialized to zero using the table-based
mechanism instead of relying on the startup code to initialize it
explicitly.

This aligns the script to the other existing ones which already
include .TFM_BSS in the zero table.

Change-Id: Iae2dd8a1f1e3a421e5162ecc8ae6cfa4cf78ec0f
Signed-off-by: Keith Packard <keithp@keithp.com>
Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>