rse: add SW mitigation for missing SAM DMA ACK in CM/DM states
In certain lifecycle states (CM/DM), the ADA DMA may not yet be active
when the CPU attempts to write to SAMICV. As a result, the SAM FSM waits
indefinitely for the `sam_config_done_trig_ack` signal, blocking further
configuration.

This change introduces a software workaround that uses the DMA350 driver
to manually generate the SAM configuration done trigger ACK. The ACK is
only triggered once and is skipped in SE state, where the ADA DMA is
expected to provision SAM correctly.

This ensures that SAM initialization completes successfully even when
ADA DMA is not yet configured.

Change-Id: If1d632a100514aa595dd55a81225220645577b9a
Signed-off-by: Amjad Ouled-Ameur <amjad.ouled-ameur@arm.com>