BL1_2: RSE: Expand BL1_2 data region to support image binding
In order to overwrite the newly computed tag when image binding is
enabled, flash driver needs to erase and reprogram the corresponding
sector. This means it needs atleast sector size buffer to backup the
data before erasing. Therefore, the BL1_2 data region size needs to be
increased to accommodate the extra data.

Signed-off-by: Maulik Patel <maulik.patel@arm.com>
Change-Id: Ic3fe10fde7b8a55922ab2cec46504630f29067fd