Revert "RSE: Remove duplicate shared measurement area clear"
We need to clear the shared_boot_measurement area on each boot,
regardless of whether or not the DMA has run to clear the persistent
boot area. Failure to do this will cause the boot measurement to fail
when the system resets and the persistent data area has been setup.

This reverts commit 999204653e48c9ec37b31791ccdcaf27d899062e.

Change-Id: If074ae25c5d823c658b932f43026408c59f99b09
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>