rse: fix improper word-wise increment of LFT OTP counter
The `lft_counter` field is defined as a `uint8_t[]`, but the increment
loop operates on 32-bit words. This caused incorrect behavior since the
array was not cast to a `uint32_t*` before word-wise access.

Fix this by explicitly casting `lft_counter` to a `volatile uint32_t*`
prior to dereferencing, ensuring proper alignment and correct counter
increment.

Change-Id: I3d4fd88a191bb709555164a9985afa5f8a536368
Signed-off-by: Amjad Ouled-Ameur <amjad.ouled-ameur@arm.com>