RSE: Require at least 4 MHU channels
Currently the RSE SFCP link HAL and MHU wrapper only require a minimum
of 2 MHU channels. With this number of channels the send and receive
functions will not work correctly as will skip the signalling step and
continue to write into non-existent channels.

Fix this by enforcing at least 4 channels, this is sufficient for the
logic to work correctly and also ensures adequate performance with the
current send/receive logic.

Change-Id: I82127a0db77989d6c57a9debb47f156037f4b8e8
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>