RSE: Setup image binding keys in BL1
This commit adds the function to setup the image binding key in BL1,
which is called during BL1 boot flow. The image binding key is stored
in a KMU slot for later use.

For AES-CCM binding, an extra slot is setup to hold the same key,
due to CryptoCell limitations.

Signed-off-by: Maulik Patel <maulik.patel@arm.com>
Change-Id: Ib51669f568f5f70b72d3c6316a83fc9831a41c03