RSE: SFCP: Add support for PSA forming of messages and replies
Add additional support for the PSA protocol handling source files to
serialized messages and deserialize replies.

This code is taken from trusted-firmware-a drivers/arm/rse commit
1d5aa939bc8d3d892e2ed9945fa50e36a1a924cc and has been modified to work
in the RSE context (mainly changing the function signatures slightly to
make the error codes compatible).

Note that as part of this patch, we also add Doxygen for the newly added
functions, existing internal functions for protocol and embed as well as
the various structure definitions.

Change-Id: Ic1ccac02686ee67b2d023ae727a9552a2f876074
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>