RSE: SFCP: Add implementation for psa_call
Add an implementation of psa_call which makes use of the SFCP
communication top-level API. This is intended to be used by a remote
processor to be able to make PSA calls to the RSE.

Note that this code is not actively compiled as it is not used by the
TF-M runtime. It is instead reference for other consumers of the SFCP
source code.

Change-Id: I4c941a3a585261356e395f1c9e395209ee33d7cb
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>