rse: mitigate AES DFA limitation by dividing CryptoCell clock
The CryptoCell AES engine cannot operate above 500 MHz when the DFA
countermeasure is enabled due to a hardware limitation. To work around
this, program the Integration Layer Control Unit (ILCU) to apply a
clock divider whenever the AES engine is selected (i.e when a valid
output address is provided).

TC platforms do not model the ILCU registers in FVP, so clock division
is disabled for those platforms.

Change-Id: I5d55252d5d64232dd243a2fea49c849f7c8da9cd
Signed-off-by: Amjad Ouled-Ameur <amjad.ouled-ameur@arm.com>