RSE: Split out ATU runtime config from BL2
The BL2 ATU configuration structure wrongly contained the definition for
the protected storage partition. Split out this definition into a new
atu_config_runtime.c source file which will be compiled only by the
runtime.

As part of this patch, remove the installing of the ATU runtime
configuration into the NS build, it should not be required.

Change-Id: If3fe2acdf85ce4ddb60c3956d805539777bcf7f6
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>