RSE: SFCP: Add sfcp_platform functions for fetching routing info
Previously, the SFCP link HAL directly called into RSE specific
functions to fetch the RSE ID and the routing tables. This patch makes
these calls more generic by moving their implementation into the SFCP
platform implementation instead.

This makes porting the SFCP software stack to other components (such as
TF-A) more straightforward as it allows the link HAL to be mostly
re-used and only the SFCP platform needs to be specific to the
component. We can do this because the core MHU driver APIs are
essentially the same across different components.

Note that as part of this patch we also add the missed Doxygen to the
sfcp_platform_get_trusted_subnets function.

Change-Id: I4acf1a7085c08441d9478f2a51b06f9db7c79a87
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>