platform: nordic_nrf: Fix platform_ns and tests builds
Update platform_ns to use correct names for startup files.
Add in required includes to work with new startup files.

Change the startup files to be linked as PRIVATE since
they don't need to be PUBLIC.

Make sure that the __Vectors symbol is retained when the
TF-M tests are being built because it is used by the
tfm_common_ns.ld linker script.

Change-Id: I34f28a5e5273819084bf7e95c8827d73472703b7
Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>