RSE: SFCP: Remove size_t cast to uint32_t
Large parts of the SFCP link HAL can be used on multiple different SW
components including those which have 64-bit architectures. Casting the
size_t to uint32_t is problematic because the size_t will occupy word
size. Therefore the top 4 bytes will not be initialised in the call to
mhu_channel_receive_device_receive and if the stack variable is not
initialised initially, we can end up with garbage values.

Fix this by using an intermediary uint32_t stack variable.

Change-Id: I0c5f3d52e54b35a4d31c563fe8b94aecef83d2cd
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>