RSE: Add support for SE-DEV software LCS
Add support for a software defined SE development life cycle state. This
can be set using the DCSU in a similar manner to the product specific
feature control OTP fields.

Upon boot, in BL1_2, we read the value in the OTP field and apply
platform specific DCU enable masks if the SE-DEV LCS is enabled.

Note that the SE_DEV control OTP field is combined with the existing
product specific feature control (PS_FC) field. This is because each of
these fields is 1 byte and therefore they must be read/written together
as the OTP reads/writes must be word size. The defined DCSU field for
this is also combined as the DCSU driver must also access these fields
together. In the case that one of the features is not present, the field
size is maintained with padding bytes.

Change-Id: I7949676c6709ff3047f602c8da6743cbc8fa6cf0
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>