RSE: SFCP: Add support to handle both message formats
Add support to the existing legacy message handling so that it can
support both the new message format and the legacy message format
simultaneously.

This is handled by keeping track of which node is using which format.
Whenever a legacy message is received from a node, we assume that node
is making use of legacy messages and send messages to that node in the
legacy message format. If we then receive a new format message, we then
update the buffer and send new format messages in response.

Change-Id: Idd4a9d4b2e59783490c24612f12a98bc2589545c
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>