Tools: Use cross compiler include path
cindex by default makes use of the host installed clang compiler to
generate the python definitions that are used in c_struct.py. These
headers may differ (and often do) from the system headers that the cross
compiler provides. Therefore make use of the cross compiler system
headers for consistentcy with the rest of the compilation.

We achieve this by fetching the cross-compiler path from the
compile_commands.json as well as the include paths and defines for the
source file being processed. This is fed into the c_struct.py class
which then queries it for its system include paths.

Change-Id: Ib2af759124ee8c99e3e4fb38c149a23832e7d712
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>