platform: nordic_nrf: Fix MPC interrupts while debugging
When a debugger with secure debug enabled is attached and performs stack
unwinding (e.g., backtrace), it may read garbage addresses from the stack.
These reads trigger MPC MEMACCERR events, disrupting debugging.

To get around this accesses to memory addresses that do not exist on
a given device are ignored as long as secure debugging is enabled.

Change-Id: Ifa4eae67bc2c25c5e27eadba7d629f120bccef60
Signed-off-by: Dag Erik Gjørvad <dag.erik.gjorvad@nordicsemi.no>