RSE: FWU: Align physical address ATU mapping
The physical address passed to the atu_rse_map_addr_to_log_addr function
must be aligned to the ATU page size. This alignment requirement is
platform specific but the standard size is 8192. This alignment
requirement is larger than the alignment requirement of GPT paritions
which causes mapping failures when reading the FWU metadata.

Update the FWU metadata parsing by aligning the mapped address down to
the ATU page boundary and then accessing the logical address at an
offset. This is similar to the address alignment in host_flash_atu.c.

The logical address provided is not aligned as that is hardcoded in the
RSE and therefore we enforce alignment of this value.

Change-Id: I3ea685a11aeff0a7f1b12c3cbd8a6632cf0349ca
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>