RSE: DCSU: Prevent reset when command fails
When the RSE receives a command over the DCSU and there is a failure,
the RSE will return the error code over the TX register but return
success to the caller. This causes the
provisioning_comms_receive_command_non_blocking function to incorrectly
think that the command has succeeded and it may then reset the system.
Prevent this by only returning a valid command code when the response
over the DCSU is success.

Change-Id: I2159217c0756502ae49d05ad97d49524c9f9e771
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>