RSE: Prevent non-endorsed ROTPK revocation
Add an additional argument to the rse_update_dm_rotpks function to
specify whether the caller is permitted to revoke the ROTPKs (this is
necessary should the current index be provisioned). Set this to false in
the case that we are performing non-endorsed DM provisioning to prevent
any ROTPKs from being revoked.

Also add Doxygen to all functions in rse_rotpk_revocation.h.

Change-Id: I30a988ad1052990c328339df1e4882e71ec07f4b
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>