RSE: Allow auth plain data to be signed by DM ROTPK
Add support for signing the authenticated plain data by the DM ROTPK.
This is used by the ROTPK revocation flow to sign DM ROTPK revocation
authenticated plain data messages.

The authentication used is passed through to the plain data handler
which determines whether or not to accept the message. Blobs can never
be signed by the DM ROTPK and a new error case is added to ensure this.

Change-Id: Iee20c822b4771131d845cc9068f46df2dcb0f8bf
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>