RSE: Allow host to export key when ROTPK is hash
In the case that a hash of a key is stored in the OTP and not a raw
public key, the key type is configured as RAW in crypto_keys.c. Because
this is not a public key type, the EXPORT_PUBLIC_KEY function cannot be
used and the host must use EXPORT_KEY instead. Allow this in the SFCP
permissions HAL.

Change-Id: Ib028907a4eac5fa8bac0de9ae117f2747cdc846d
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>