RSE: Tests: Add auth plain and ROTPK revocation tests
Refactor the existing blob generator and signing functions to allow the
testing of the authenticated plain data format. We add standard tests
for this new message format, similar to those for the blob.

Also add new tests for the ROTPK revocation authenticated plain data.
These tests ensure that the authenticated plain data is only accepted
should the correct ROTPK be used to sign it. They also verify that the
message contents are parsed correctly.

Change-Id: Ic95e9c615f991cb4331398f08e43ec0c0cfdca41
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>